backend/requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==6.0.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.1
requests
